更改頁面

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace shopping_mall\Http\Controllers;
 use Illuminate\Http\Request;
 use shopping_mall\Http\Requests;
 use shopping_mall\Models\Product;
+use shopping_mall\Models\Product_model;
 use shopping_mall\Models\User;
 use shopping_mall\Library\Cart;
 use Illuminate\Support\Facades\Auth;
@@ -15,17 +16,12 @@ class ProductController extends Controller
     private $product_model;
 
     public function __construct(){
-        $this->product_model = new Product();
+        $this->product_model = new Product_model();
     }
 
-    public function getIndex(){
-        $this->data['product'] =  $this->product_model->take(9)->get();
-        return view('shop/index', $this->data);
 
-    }
-
-    public function getAddToCart(Request $request, $id){
-        $product = Product::find($id);
+    public function postCart(Request $request){
+        $product = Product::find($request->input('id'));
         $oldcart = $request->session()->has('cart') ? $request->session()->get('cart') : null;
         $cart = new Cart($oldcart);
         $cart->add($product, $request->input('qty'));
@@ -33,26 +29,60 @@ class ProductController extends Controller
         return json_encode($cart);
     }
 
-    public function getDeleteToCart(Request $request, $id){
-        $oldcart = $request->session()->has('cart') ? $request->session()->get('cart') : null;
-        $cart = new Cart($oldcart);
-        $cart->delete($id);
-        $request->session()->put('cart', $cart);
+    public function deleteCart(Request $request){
+        if($id = $request->input('id')){
+            $oldcart = $request->session()->has('cart') ? $request->session()->get('cart') : null;
+            $cart = new Cart($oldcart);
+            $cart->delete($id);
+            $request->session()->put('cart', $cart);
+        }else{
+            $request->session()->forget('cart');
+        }
+
     }
 
-    public function getPage(Request $request, $type){
-        $this->data['product'] = $this->product_model->where('type', '=', $type)->take(9)->get();
+    public function putCart(Request $request){
+        $id = $request->input('id');
+        $number = $request->input('number');
+        $cart = new Cart($oldcart);
+        $cart->put($id, $number);
+        $request->session()->put('cart', $cart);
+        return route('sale.index');
+    }
+
+    public function getPage(Request $request){
+
+        $count = $request->input('page') ? $request->input('page') * 9 : 0;
+        $this->data['condition'] = '';
+        if($type = $request->input('type')){
+            $this->data['condition'] = 'type=' . $type;
+        };
+
+        if($search = $request->input('search')){
+            $this->data['condition'] = $this->data['condition'] . '&search=' . $search;
+        }
+        $this->data['page'] = $count/9;
+        $this->data['total'] = $this->product_model->getNumber($type, $search);
+        $this->data['product'] = $this->product_model->getPage($type, $search, $count);
         return view('shop/index', $this->data);
     }
 
-    public function postSearch(Request $request){
-
-    }
 
     public function test_cart_add(Request $request){
-       /*$user = User::first();
-       print_r($user);*/
-       $request->session()->forget('cart');
+        $count = $request->input('page') ? $request->input('page') * 9 : 0;
+        $this->data['condition'] = '';
+        if($type = $request->input('type')){
+            $this->data['condition'] = 'type=' . $type;
+            $this->product_model = $this->product_model->where('type', '=', $type);
+        };
+
+        if($search = $request->input('search')){
+            $this->product_model = $this->product_model->where('title', 'like', '%' .$search. '%');
+            $this->data['condition'] = $this->data['condition'] . '&' . $search;
+        }
+
+        $product_model = new Product_model();
+        echo $product_model->getNumber($type, $search, $count);
 
 
     }
